// src/store/product/types.ts
export enum ActionTypes {
  GET_COUNTER = "GET_COUNTER",
  GET_PRODUCTS = "GET_PRODUCTS",
  GET_CATEGORIES = "GET_CATEGORIES",
  GET_PRODUCTS_BY_CATEGORY = "GET_PRODUCTS_BY_CATEGORY",
}

export enum MutationTypes {
  SET_COUNTER = "SET_COUNTER",
  SET_PRODUCTS = "SET_PRODUCTS",
  SET_CATEGORIES = "SET_CATEGORIES",
  SET_LOADING = "SET_LOADING",
}